splitted utiils file and added training run endpoint
